Moves encrypt methods into Encrypter module
