+++ lib/encrypter.rb
@@ -1,3 +1,15 @@
-class Encrypter
+module Encrypter
+  def encrypt_hash(encryption,key,offset)
+    {encryption: encryption,
+    key: key,
+    date: offset}
+  end
 
+  def encrypt(message,key,offset)
+    message.downcase!
+    @shifts.create_hashes(key,offset) #could this go in create_shifted_arrays?
+    @shifts.create_shifted_arrays(key,offset)
+    shift_all(message)
+    encrypt_hash(message,key,offset)
+  end
 end
