+++ lib/enigma.rb
@@ -1,11 +1,13 @@
 require "./lib/shifts"
 require "./lib/letter_shift"
 require "./lib/letter_shift_back"
+require "./lib/encrypter"
 require "./lib/decrypter"
 
 class Enigma
   include LetterShift
   include LetterShiftBack
+  include Encrypter
   include Decrypter
 
   attr_reader :shifts,
@@ -21,19 +23,5 @@ class Enigma
 
   def char_index(message)
     message.chars.map {|char| @alphabet.index(char)}
-  end
-
-  def encrypt_hash(encryption,key,offset)
-    {encryption: encryption,
-    key: key,
-    date: offset}
-  end
-
-  def encrypt(message,key,offset)
-    message.downcase!
-    @shifts.create_hashes(key,offset) #could this go in create_shifted_arrays?
-    @shifts.create_shifted_arrays(key,offset)
-    shift_all(message)
-    encrypt_hash(message,key,offset)
   end
 end
