+++ test/enigma_test.rb
@@ -42,41 +42,4 @@ class EnigmaTest < MiniTest::Test
   def test_message_chars_indices_in_array
     assert_equal [3,20,2,10,18], @enigma.char_index("ducks") #this requires an attr_writer...
   end
-
-  def test_encrypt_hash_method_returns_hash
-    expected = {encryption: "message",
-                key: "12345",
-                date: "032489"}
-    assert_equal expected, @enigma.encrypt_hash("message","12345","032489")
-  end
-
-  def test_encrypt_method_creates_encrypted_message
-    expected = {encryption: "urlci",
-                key: "12345",
-                date: "032489"}
-    assert_equal expected, @enigma.encrypt("Ducks", "12345", "032489")
-  end
-
-  def test_encrypt_method_creates_encrypted_message_with_capitals
-    expected = {encryption: "urlci",
-                key: "12345",
-                date: "032489"}
-    assert_equal expected, @enigma.encrypt("DUCKS", "12345", "032489")
-  end
-
-  def test_encrypt_method_ignores_numbers_or_punctuation
-    expected = {encryption: "urlci.",
-                key: "12345",
-                date: "032489"}
-    assert_equal expected, @enigma.encrypt("ducks.", "12345", "032489")
-    expected = {encryption: "urlci2",
-                key: "12345",
-                date: "032489"}
-    assert_equal expected, @enigma.encrypt("ducks2", "12345", "032489")
-  end
-
-  def test_key_not_given_creates_random_number_encrypted_message
-    skip
-    assert_equal expected, @enigma.encrypt("ducks", "12345", "032489")
-  end
 end
